Semana-7/ROJAS_BURBANO_Centro_Estético/requirements.txt
anyio==4.10.0
colorama==0.4.6
idna==3.10
iniconfig==2.1.0
memory-profiler==0.61.0
packaging==25.0
pluggy==1.6.0
prometheus-fastapi-instrumentator==7.1.0
prometheus_client==0.23.1
psutil==7.1.0
Pygments==2.19.2
pytest==8.4.2
redis==6.4.0
sniffio==1.3.1
starlette==0.48.0
